Separate april_camera into utilities and examples
