+++ april_vision/helpers/sender.py
@@ -1,0 +1,84 @@
+import base64
+from threading import Thread
+from typing import List, Optional, Callable
+
+import cv2
+import numpy as np
+from numpy.typing import NDArray
+
+from ..marker import Marker
+from ..utils import Frame
+from ..vision import Processor
+
+
+class Base64Sender:
+    def __init__(
+        self,
+        publish_callback: Callable[[str, bytes], None],
+        *,
+        annotated: bool = True,
+        threaded: bool = True,
+        aruco_orientation: bool = True,
+    ):
+        self._publish_callback = publish_callback
+        self.use_threads = threaded
+        self.send_annotated = annotated
+        self._processor = Processor(aruco_orientation=aruco_orientation)
+
+    def annotated_frame_hook(self, frame: Frame, markers: List[Marker]) -> None:
+        """
+
+        The image data is a base64 JPEG.
+        """
+        copied_frame = Frame(
+            np.array(frame.grey_frame, copy=True),
+            np.array(frame.colour_frame, copy=True),
+        )
+        if self.send_annotated:
+            output_frame = self._processor._annotate(copied_frame, markers)
+        else:
+            output_frame = copied_frame
+
+        if self.use_threads:
+            thread = Thread(
+                name="Image send",
+                target=self.encode_and_send,
+                args=(output_frame.colour_frame,),
+                daemon=True,
+            )
+            thread.start()
+        else:
+            self.encode_and_send(output_frame.colour_frame)
+
+    def encode_and_send(self, frame: NDArray[np.uint8]) -> None:
+        """
+        Handle converting a frame to a base64 bytestring and sending it using
+        the publish callback.
+
+        Can be run as a thread target.
+        """
+        if self._publish_callback is None:
+            return
+
+        encoded_frame = self.base64_encode_frame(frame)
+        if encoded_frame is None:
+            return
+
+        encoded_frame = b'data:image/jpeg;base64, ' + encoded_frame
+
+        # The message is sent in a background thread
+        try:
+            self._publish_callback(
+                'camera/annotated' if self.send_annotated else 'camera/image',
+                encoded_frame,
+            )
+        except ValueError:
+            pass
+
+    def base64_encode_frame(self, frame: NDArray[np.uint8]) -> Optional[bytes]:
+        """Convert image frame to base64 bytestring."""
+        ret, image = cv2.imencode('.jpg', frame)
+        if not ret:
+            return None
+        image_bytes = image.tobytes()
+        return base64.b64encode(image_bytes)
