--- april_vision/j5/__init__.py
@@ -1,7 +1,0 @@
-"""j5 integration for april_vision."""
-from .april_camera import AprilCameraBoard, AprilTagHardwareBackend
-
-__all__ = [
-    'AprilCameraBoard',
-    'AprilTagHardwareBackend',
-]
