+++ setup.cfg
@@ -27,8 +27,6 @@ dev =
     build
     types-Pillow
     types-tabulate
-j5 =
-    j5 >=1, <2
 cli =
     opencv-python >=4, <5
     Pillow >=9.4.0, <10
